Update 2b script to analyse by region
Rename to analyse.
Update script to analyse each region separately.
Update script to use prop_suitability to calculate the proportion of hectare that is suitable.
